refactor(cli): moves finish setup to own function
